removed visualizations and moved relevant functions to analysis.py
